feat: panning is done via JS now

+++ src/FlowState/Components/FlowNodeBase.cs
@@ -8,10 +8,10 @@ namespace FlowState.Components;
 public abstract class FlowNodeBase : ComponentBase, IDisposable
 {
     [Parameter]
-    public int X { get; set; }
+    public double X { get; set; }
 
     [Parameter]
-    public int Y { get; set; }
+    public double Y { get; set; }
 
     public string Id { get; } = Guid.NewGuid().ToString();
     public abstract ValueTask ExecuteAsync();
